LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/traverseTo-navigates-multiple-iframes-expected.txt



FAIL entries() should be correct after a traversal that navigates multiple browsing contexts assert_equals: expected 2 but got 1
